+++ lib/features/activepair/domain/usecase/get_entity_pairs_summary_use_case.dart
@@ -1,0 +1,15 @@
+import 'package:zachranobed/features/activepair/domain/model/entity_pairs_summary.dart';
+import 'package:zachranobed/features/activepair/domain/repository/entity_pairs_repository.dart';
+import 'package:zachranobed/models/user_data.dart';
+
+/// A use case for retrieving entity pairs summary information.
+class GetEntityPairsSummaryUseCase {
+  final EntityPairsRepository _repository;
+
+  GetEntityPairsSummaryUseCase(this._repository);
+
+  /// Fetches available pairs summary information for the [user].
+  Future<EntityPairsSummary> invoke(UserData user) {
+    return _repository.getEntityPairsSummary(user: user);
+  }
+}