+++ lib/main.dart
@@ -5,6 +5,7 @@ import 'package:intl/date_symbol_data_local.dart';
 import 'package:zachranobed/common/di/common_dependency_container.dart';
 import 'package:zachranobed/common/logger/zo_logger.dart';
 import 'package:zachranobed/common/firebase/firebase_helper.dart';
+import 'package:zachranobed/features/activepair/di/active_pair_dependency_container.dart';
 import 'package:zachranobed/features/appConfiguration/app_configuration.dart';
 import 'package:zachranobed/features/appConfiguration/mapper/app_configuration_mapper.dart';
 import 'package:zachranobed/features/appTerms/di/app_terms_dependency_container.dart';
@@ -34,6 +35,7 @@ void main() async {
   FoodBoxDependencyContainer.setup();
   OfferedFoodDependencyContainer.setup();
   MenuDependencyContainer.setup();
+  ActivePairDependencyContainer.setup();
 
   ZOLogger.init();