+++ lib/models/charity.dart
@@ -1,9 +1,7 @@
+import 'package:zachranobed/models/entity_pair.dart';
 import 'package:zachranobed/models/user_data.dart';
 
 class Charity extends UserData {
-  final List<String> donorIds;
-  final List<String> carrierIds;
-
   Charity({
     required super.entityId,
     required super.email,
@@ -11,7 +9,21 @@ class Charity extends UserData {
     required super.establishmentId,
     required super.organization,
     required super.lastAcceptedAppTermsVersion,
-    required this.donorIds,
-    required this.carrierIds,
+    required super.activePair,
+    required super.hasMultiplePairs,
   });
+
+  @override
+  UserData copyWith({required EntityPair activePair}) {
+    return Charity(
+      entityId: entityId,
+      email: email,
+      establishmentName: establishmentName,
+      establishmentId: establishmentId,
+      organization: organization,
+      lastAcceptedAppTermsVersion: lastAcceptedAppTermsVersion,
+      hasMultiplePairs: hasMultiplePairs,
+      activePair: activePair,
+    );
+  }
 }