+++ lib/models/user_data.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:zachranobed/models/entity_pair.dart';
 
 abstract class UserData extends ChangeNotifier {
   final String entityId;
@@ -7,6 +8,8 @@ abstract class UserData extends ChangeNotifier {
   final String establishmentId;
   final String organization;
   final int? lastAcceptedAppTermsVersion;
+  final EntityPair activePair;
+  final bool hasMultiplePairs;
 
   UserData({
     required this.entityId,
@@ -14,7 +17,9 @@ abstract class UserData extends ChangeNotifier {
     required this.establishmentName,
     required this.establishmentId,
     required this.organization,
-    required this.lastAcceptedAppTermsVersion
+    required this.lastAcceptedAppTermsVersion,
+    required this.activePair,
+    required this.hasMultiplePairs,
   });
 
   String get debugInfo {
@@ -24,7 +29,12 @@ abstract class UserData extends ChangeNotifier {
       Establishment: $establishmentName,
       ID: $establishmentId,
       Organization: $organization,
-      Last accepted app terms version: $lastAcceptedAppTermsVersion
+      Last accepted app terms version: $lastAcceptedAppTermsVersion,
+      Active pair: ${activePair.donorId} <-> ${activePair.recipientId},
+      Has multiple pairs: $hasMultiplePairs
     ''';
   }
+
+  /// Creates a copy of this [UserData] object with a replaced [activePair].
+  UserData copyWith({required EntityPair activePair});
 }