+++ lib/services/di/services_dependency_container.dart
@@ -1,4 +1,5 @@
 import 'package:get_it/get_it.dart';
+import 'package:zachranobed/common/prefs/app_preferences.dart';
 import 'package:zachranobed/routes/app_router.dart';
 import 'package:zachranobed/services/auth_service.dart';
 import 'package:zachranobed/services/carrier_service.dart';
@@ -22,6 +23,7 @@ class ServicesDependencyContainer {
       AuthService(
         GetIt.I<EntityService>(),
         GetIt.I<EntityPairService>(),
+        GetIt.I<AppPreferences>(),
       ),
     );
     GetIt.I.registerSingleton(ConfigurationService());